Move top level elements to separate namespace
