+++ ClosedXML.IO.CodeGen/Model/Schema.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using ClosedXML.IO.CodeGen.Model.TopLevel;
+using System.Collections.Generic;
 
 namespace ClosedXML.IO.CodeGen.Model;
 
@@ -13,7 +14,7 @@ public class Schema
     public List<ImportElement> Imports = [];
 
     /// <summary>
-    /// One of <see cref="AttributeGroupDefinition"/>, <see cref="ComplexTypeBase"/>, <see cref="ElementDefinition"/>.
+    /// One of <see cref="AttributeGroupDefinition"/>, <see cref="ComplexType"/>, <see cref="ElementDefinition"/>.
     /// </summary>
     public List<object> Entries = [];
 }
